Better handling of distraction mode state
Previously, the distraction mode was active when the sidebar wasn't visible, so the addon didn't work well for users that have their sidebar hidden.
Now, the current state is saved inside the window's settings, so the state is independent of the user's configuration and yet gets saved between sessions. The long name of the setting was chosen on purpose to prevent possible name collisions.